// components/common/modals/modal-sell-v2/type.d.ts
export type Data = {
    saleType: number
    fixedPrice: string
    startPrice: string
    startTime: number
    closeTime: number
    basePrice: number
    chainId: number
    error: {
        fixedPrice: string
        closeTime: string
        startTime: string
        startPrice: string
    }
}